move violation endpoint tests to own file
